IMplementation of recovering previous entry server ID's.

+++ server/play/app/models/PictureCollection.java
@@ -1,13 +1,16 @@
 package models;
 
 import java.util.*;
+
 import javax.persistence.*;
 
 import com.avaje.ebean.Model;
+
 import play.data.format.*;
 import play.data.validation.*;
 import play.Logger;
 import modules.AmazonHelper;
+
 import java.io.File;
 
 import com.avaje.ebean.*;
@@ -15,14 +18,14 @@ import com.avaje.ebean.*;
 /**
  * Groups of pictures entity managed by Ebean
  */
-@Entity 
+@Entity
 public class PictureCollection extends Model {
 
     private static final long serialVersionUID = 1L;
 
-	@Id
+    @Id
     public Long id;
-    
+
     @Constraints.Required
     public Long collection_id;
 
@@ -32,9 +35,11 @@ public class PictureCollection extends Model {
     @Constraints.Required
     public String note;
 
-    public static Finder<Long,PictureCollection> find = new Finder<Long,PictureCollection>(PictureCollection.class);
+    public static Finder<Long, PictureCollection> find = new Finder<Long, PictureCollection>(PictureCollection.class);
 
-    public static List<PictureCollection> list() { return find.all(); }
+    public static List<PictureCollection> list() {
+        return find.all();
+    }
 
     public static List<PictureCollection> findByCollectionId(long collection_id) {
         return find.where()
@@ -47,8 +52,10 @@ public class PictureCollection extends Model {
                 .eq("collection_id", collection_id)
                 .findList();
         for (PictureCollection item : items) {
-            File file = amazonHelper.getLocalFile(item.picture);
-            file.delete();
+            if (amazonHelper != null) {
+                File file = amazonHelper.getLocalFile(item.picture);
+                file.delete();
+            }
             item.delete();
         }
     }
